Refactor delete annotation and comment logger and generator
