+++ webserver/src/activity-logger/generator/annotation-comment/annotation-generator.js
@@ -1,0 +1,445 @@
+import { v4 as uuidv4 } from "uuid";
+import {
+  createContext,
+  createMetadata,
+  createUser,
+  createVerb,
+} from "../util/generator-util";
+import {
+  createAnnotationCommentMaterialObject,
+  createAnnotationCommentObject,
+  createAnnotationCommentMaterialResultObject,
+  createAnnotationCommentResultObject,
+} from "./utils";
+
+const platform = "CourseMapper";
+const language = "en-US";
+
+export const generateCreateAnnotationActivity = (req) => {
+  const metadata = createMetadata();
+  return {
+    ...metadata,
+    actor: createUser(req),
+    verb: createVerb(
+      "http://risc-inc.com/annotator/verbs/annotated",
+      "annotated",
+    ),
+    object: createAnnotationCommentMaterialObject(req),
+    result: createAnnotationCommentMaterialResultObject(req, "annotation"),
+    context: createContext(),
+  };
+};
+
+export const generateDeleteAnnotationActivity = (req) => {
+  const metadata = createMetadata();
+  return {
+    ...metadata,
+    actor: createUser(req),
+    verb: createVerb("http://activitystrea.ms/schema/1.0/delete", "deleted"),
+    object: createAnnotationCommentObject(req, "annotation"),
+    result: createAnnotationCommentResultObject(req, "annotation"),
+    context: createContext(),
+  };
+};
+
+export const generateLikeAnnotationActivity = (user, annotation, origin) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://activitystrea.ms/schema/1.0/like",
+      display: {
+        [language]: "liked",
+      },
+    },
+    object: {
+      objectType: "Activity",
+      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/annotation",
+        name: {
+          [language]:
+            "Annotation:" +
+            annotation.content.slice(0, 50) +
+            (annotation.content.length > 50 ? " ..." : ""),
+        },
+        description: {
+          [language]: annotation.content,
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: annotation._id,
+            material_id: annotation.materialId,
+            channel_id: annotation.channelId,
+            topic_id: annotation.topicId,
+            course_id: annotation.courseId,
+            content: annotation.content,
+            type: annotation.type,
+            tool: annotation.tool,
+            location: annotation.location,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          location: annotation.location,
+        },
+      },
+    },
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
+
+export const getAnnotationUnlikeStatement = (user, annotation, origin) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://activitystrea.ms/schema/1.0/unlike",
+      display: {
+        [language]: "unliked",
+      },
+    },
+    object: {
+      objectType: "Activity",
+      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/annotation",
+        name: {
+          [language]:
+            "Annotation:" +
+            annotation.content.slice(0, 50) +
+            (annotation.content.length > 50 ? " ..." : ""),
+        },
+        description: {
+          [language]: annotation.content,
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: annotation._id,
+            material_id: annotation.materialId,
+            channel_id: annotation.channelId,
+            topic_id: annotation.topicId,
+            course_id: annotation.courseId,
+            content: annotation.content,
+            type: annotation.type,
+            tool: annotation.tool,
+            location: annotation.location,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          location: annotation.location,
+        },
+      },
+    },
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
+
+export const generateDislikeAnnotationActivity = (user, annotation, origin) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://activitystrea.ms/schema/1.0/dislike",
+      display: {
+        [language]: "disliked",
+      },
+    },
+    object: {
+      objectType: "Activity",
+      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/annotation",
+        name: {
+          [language]:
+            "Annotation:" +
+            annotation.content.slice(0, 50) +
+            (annotation.content.length > 50 ? " ..." : ""),
+        },
+        description: {
+          [language]: annotation.content,
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: annotation._id,
+            material_id: annotation.materialId,
+            channel_id: annotation.channelId,
+            topic_id: annotation.topicId,
+            course_id: annotation.courseId,
+            content: annotation.content,
+            type: annotation.type,
+            tool: annotation.tool,
+            location: annotation.location,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          location: annotation.location,
+        },
+      },
+    },
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
+
+export const generateUndislikeAnnotationActivity = (
+  user,
+  annotation,
+  origin,
+) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://www.CourseMapper.de/verbs/undisliked",
+      display: {
+        [language]: "un-disliked",
+      },
+    },
+    object: {
+      objectType: "Activity",
+      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/annotation",
+        name: {
+          [language]:
+            "Annotation:" +
+            annotation.content.slice(0, 50) +
+            (annotation.content.length > 50 ? " ..." : ""),
+        },
+        description: {
+          [language]: annotation.content,
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: annotation._id,
+            material_id: annotation.materialId,
+            channel_id: annotation.channelId,
+            topic_id: annotation.topicId,
+            course_id: annotation.courseId,
+            content: annotation.content,
+            type: annotation.type,
+            tool: annotation.tool,
+            location: annotation.location,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          location: annotation.location,
+        },
+      },
+    },
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
+
+export const generateEditAnnotationActivity = (
+  user,
+  newAnnotation,
+  oldAnnotation,
+  origin,
+) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://curatr3.com/define/verb/edited",
+      display: {
+        [language]: "edited",
+      },
+    },
+    object: {
+      objectType: "Activity",
+      id: `${origin}/activity/course/${oldAnnotation.courseId}/topic/${oldAnnotation.topicId}/channel/${oldAnnotation.channelId}/material/${oldAnnotation.materialId}/annotation/${oldAnnotation._id}`,
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/annotation",
+        name: {
+          [language]:
+            "Annotation:" +
+            oldAnnotation.content.slice(0, 50) +
+            (oldAnnotation.content.length > 50 ? " ..." : ""),
+        },
+        description: {
+          [language]: oldAnnotation.content,
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: oldAnnotation._id,
+            material_id: oldAnnotation.materialId,
+            channel_id: oldAnnotation.channelId,
+            topic_id: oldAnnotation.topicId,
+            course_id: oldAnnotation.courseId,
+            content: oldAnnotation.content,
+            type: oldAnnotation.type,
+            tool: oldAnnotation.tool,
+            location: oldAnnotation.location,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          content: newAnnotation.content,
+          type: newAnnotation.type,
+          tool: newAnnotation.tool,
+          location: newAnnotation.location,
+        },
+      },
+    },
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
+
+export const generateAddMentionStatement = (user, annotation, origin) => {
+  const userId = user._id.toString();
+  const userFullname = `${user.firstname} ${user.lastname}`;
+  return {
+    id: uuidv4(),
+    timestamp: new Date(),
+    actor: {
+      objectType: "Agent",
+      name: userFullname,
+      mbox: user.mbox,
+      mbox_sha1sum: user.mbox_sha1sum,
+      account: {
+        homePage: origin,
+        name: userId,
+      },
+    },
+    verb: {
+      id: "http://id.tincanapi.com/verb/mentioned",
+      display: {
+        [language]: "mentioned",
+      },
+    },
+    object: {
+      objectType: "User",
+
+      definition: {
+        type: "http://www.CourseMapper.de/activityType/you",
+        name: {
+          [language]: "",
+        },
+        extensions: {
+          "http://www.CourseMapper.de/extensions/annotation": {
+            id: annotation._id,
+            material_id: annotation.materialId,
+            channel_id: annotation.channelId,
+            topic_id: annotation.topicId,
+            course_id: annotation.courseId,
+            content: annotation.content,
+          },
+        },
+      },
+    },
+    result: {
+      extensions: {
+        "http://www.CourseMapper.de/extensions/annotation": {
+          id: annotation._id,
+          material_id: annotation.materialId,
+          channel_id: annotation.channelId,
+          topic_id: annotation.topicId,
+          course_id: annotation.courseId,
+          content: annotation.content,
+          type: annotation.type,
+          tool: annotation.tool,
+          location: annotation.location,
+        },
+      },
+    },
+
+    context: {
+      platform: platform,
+      language: language,
+    },
+  };
+};
