--- webserver/src/activity-logger/generator/annotation-generator.js
@@ -1,1006 +1,0 @@
-import { v4 as uuidv4 } from "uuid";
-
-const platform = "CourseMapper";
-const language = "en-US";
-
-export const getAnnotationCreationStatement = (
-  user,
-  annotation,
-  material,
-  origin
-) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://risc-inc.com/annotator/verbs/annotated",
-      display: {
-        [language]: "annotated",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${material.courseId}/topic/${material.topicId}/channel/${material.channelId}/material/${material._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/material",
-        name: {
-          [language]: material.name,
-        },
-        description: {
-          [language]: material.description,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/material": {
-            id: material._id,
-            type: material.type,
-            channel_id: material.channelId,
-            topic_id: material.topicId,
-            course_id: material.courseId,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          id: annotation._id,
-          material_id: annotation.materialId,
-          channel_id: annotation.channelId,
-          topic_id: annotation.topicId,
-          course_id: annotation.courseId,
-          content: annotation.content,
-          type: annotation.type,
-          tool: annotation.tool,
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentCreationStatement = (
-  user,
-  annotation,
-  material,
-  origin
-) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://adlnet.gov/expapi/verbs/commented",
-      display: {
-        [language]: "commented",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${material.courseId}/topic/${material.topicId}/channel/${material.channelId}/material/${material._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/material",
-        name: {
-          [language]: material.name,
-        },
-        description: {
-          [language]: material.description,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/material": {
-            id: material._id,
-            type: material.type,
-            channel_id: material.channelId,
-            topic_id: material.topicId,
-            course_id: material.courseId,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/comment": {
-          id: annotation._id,
-          material_id: annotation.materialId,
-          channel_id: annotation.channelId,
-          topic_id: annotation.topicId,
-          course_id: annotation.courseId,
-          content: annotation.content,
-          type: annotation.type,
-          tool: annotation.tool,
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotaionDeletionStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/delete",
-      display: {
-        [language]: "deleted",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentDeletionStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/delete",
-      display: {
-        [language]: "deleted",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/comment/${annotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotationLikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/like",
-      display: {
-        [language]: "liked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentLikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/like",
-      display: {
-        [language]: "liked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/comment/${annotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotationUnlikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/unlike",
-      display: {
-        [language]: "unliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentUnlikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/unlike",
-      display: {
-        [language]: "unliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/comment/${annotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/comment": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotationDislikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/dislike",
-      display: {
-        [language]: "disliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentDislikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://activitystrea.ms/schema/1.0/dislike",
-      display: {
-        [language]: "disliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/comment/${annotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/comment": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotationUndislikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://www.CourseMapper.de/verbs/undisliked",
-      display: {
-        [language]: "un-disliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/annotation/${annotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentUndislikeStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://www.CourseMapper.de/verbs/undisliked",
-      display: {
-        [language]: "un-disliked",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${annotation.courseId}/topic/${annotation.topicId}/channel/${annotation.channelId}/material/${annotation.materialId}/comment/${annotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            annotation.content.slice(0, 50) +
-            (annotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: annotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-            type: annotation.type,
-            tool: annotation.tool,
-            location: annotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          location: annotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getAnnotationEditStatement = (
-  user,
-  newAnnotation,
-  oldAnnotation,
-  origin
-) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://curatr3.com/define/verb/edited",
-      display: {
-        [language]: "edited",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${oldAnnotation.courseId}/topic/${oldAnnotation.topicId}/channel/${oldAnnotation.channelId}/material/${oldAnnotation.materialId}/annotation/${oldAnnotation._id}`,
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/annotation",
-        name: {
-          [language]:
-            "Annotation:" +
-            oldAnnotation.content.slice(0, 50) +
-            (oldAnnotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: oldAnnotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: oldAnnotation._id,
-            material_id: oldAnnotation.materialId,
-            channel_id: oldAnnotation.channelId,
-            topic_id: oldAnnotation.topicId,
-            course_id: oldAnnotation.courseId,
-            content: oldAnnotation.content,
-            type: oldAnnotation.type,
-            tool: oldAnnotation.tool,
-            location: oldAnnotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          content: newAnnotation.content,
-          type: newAnnotation.type,
-          tool: newAnnotation.tool,
-          location: newAnnotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getCommentEditStatement = (
-  user,
-  newAnnotation,
-  oldAnnotation,
-  origin
-) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://curatr3.com/define/verb/edited",
-      display: {
-        [language]: "edited",
-      },
-    },
-    object: {
-      objectType: "Activity",
-      id: `${origin}/activity/course/${oldAnnotation.courseId}/topic/${oldAnnotation.topicId}/channel/${oldAnnotation.channelId}/material/${oldAnnotation.materialId}/comment/${oldAnnotation._id}`,
-      definition: {
-        type: "http://activitystrea.ms/schema/1.0/comment",
-        name: {
-          [language]:
-            "Comment:" +
-            oldAnnotation.content.slice(0, 50) +
-            (oldAnnotation.content.length > 50 ? " ..." : ""),
-        },
-        description: {
-          [language]: oldAnnotation.content,
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/comment": {
-            id: oldAnnotation._id,
-            material_id: oldAnnotation.materialId,
-            channel_id: oldAnnotation.channelId,
-            topic_id: oldAnnotation.topicId,
-            course_id: oldAnnotation.courseId,
-            content: oldAnnotation.content,
-            type: oldAnnotation.type,
-            tool: oldAnnotation.tool,
-            location: oldAnnotation.location,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/comment": {
-          content: newAnnotation.content,
-          type: newAnnotation.type,
-          tool: newAnnotation.tool,
-          location: newAnnotation.location,
-        },
-      },
-    },
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
-
-export const getNewMentionCreationStatement = (user, annotation, origin) => {
-  const userId = user._id.toString();
-  const userFullname = `${user.firstname} ${user.lastname}`;
-  return {
-    id: uuidv4(),
-    timestamp: new Date(),
-    actor: {
-      objectType: "Agent",
-      name: userFullname,
-      mbox: user.mbox,
-      mbox_sha1sum: user.mbox_sha1sum,
-      account: {
-        homePage: origin,
-        name: userId,
-      },
-    },
-    verb: {
-      id: "http://id.tincanapi.com/verb/mentioned",
-      display: {
-        [language]: "mentioned",
-      },
-    },
-    object: {
-      objectType: "User",
-
-      definition: {
-        type: "http://www.CourseMapper.de/activityType/you",
-        name: {
-          [language]: "",
-        },
-        extensions: {
-          "http://www.CourseMapper.de/extensions/annotation": {
-            id: annotation._id,
-            material_id: annotation.materialId,
-            channel_id: annotation.channelId,
-            topic_id: annotation.topicId,
-            course_id: annotation.courseId,
-            content: annotation.content,
-          },
-        },
-      },
-    },
-    result: {
-      extensions: {
-        "http://www.CourseMapper.de/extensions/annotation": {
-          id: annotation._id,
-          material_id: annotation.materialId,
-          channel_id: annotation.channelId,
-          topic_id: annotation.topicId,
-          course_id: annotation.courseId,
-          content: annotation.content,
-          type: annotation.type,
-          tool: annotation.tool,
-          location: annotation.location,
-        },
-      },
-    },
-
-    context: {
-      platform: platform,
-      language: language,
-    },
-  };
-};
