+++ webserver/src/routes/annotation.routes.js
@@ -1,6 +1,6 @@
 const { authJwt, notifications } = require("../middlewares");
 const controller = require("../controllers/annotation.controller");
-const logger = require("../activity-logger/logger/annotation-logger");
+const logger = require("../activity-logger/logger/annotation-comment-logger");
 const notificationsController = require("../controllers/notification.controller");
 
 module.exports = function (app) {
@@ -15,11 +15,11 @@ module.exports = function (app) {
     "/api/courses/:courseId/materials/:materialId/annotation",
     [authJwt.verifyToken, authJwt.isEnrolled],
     controller.newAnnotation,
-    logger.newAnnotation,
+    logger.createAnnotationLogger,
     notificationsController.followAnnotation,
     notifications.newAnnotationNotificationUsersCalculate,
     notifications.populateUserNotification,
-    logger.newMention,
+    logger.addMentionLogger,
     notifications.newMentionNotificationUsersCalculate,
     notifications.populateUserNotification,
   );
@@ -30,7 +30,7 @@ module.exports = function (app) {
     "/api/courses/:courseId/annotations/:annotationId",
     [authJwt.verifyToken, authJwt.isEnrolled],
     controller.deleteAnnotation,
-    logger.deleteAnnotation,
+    logger.deleteAnnotationLogger,
     notifications.calculateUsersFollowingAnnotation,
     notifications.populateUserNotification,
   );
@@ -53,7 +53,7 @@ module.exports = function (app) {
     "/api/courses/:courseId/annotations/:annotationId/like",
     [authJwt.verifyToken, authJwt.isEnrolled],
     controller.likeAnnotation,
-    logger.likeAnnotation,
+    logger.likeAnnotationLogger,
     notifications.LikesDislikesAnnotationNotificationUsers,
     notifications.populateUserNotification,
   );
@@ -65,7 +65,7 @@ module.exports = function (app) {
     "/api/courses/:courseId/annotations/:annotationId/dislike",
     [authJwt.verifyToken, authJwt.isEnrolled],
     controller.dislikeAnnotation,
-    logger.dislikeAnnotation,
+    logger.dislikeAnnotationLogger,
     notifications.LikesDislikesAnnotationNotificationUsers,
     notifications.populateUserNotification,
   );
